+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New Features
 
 * Added DiffViewer modes or modified images: Side-By-Side, Swipe, and Difference.
+* Add Conflict Merge tool to DiffViewer for resolving conflicted files.
 
 ## [v0.1.7](https://github.com/brendan-duncan/ugit/releases/tag/v0.1.7) (February 20, 2026)